# add WEBGL_FLUSH_THRESHOLD flag