We have notice that in Android, GL commands are not flushed periodically as desktop chrome (every 1ms), it is set to be flushed at the end of the javascript task, this means no GL execution until the end of the model execution is done is javascript.

Created a flag for trigger manual gl flush, I have observed significant performance gain on mobilenet with Andorid devices (flush every 0.8ms):
One+ 8 phone:  48ms -> 20ms
Pixel 5: 64 -> 40ms
pixel 4: 54ms -> 24ms


To see the logs from the Cloud Build CI, please join either our [discussion](https://groups.google.com/a/tensorflow.org/forum/#!forum/tfjs) or [announcement](https://groups.google.com/a/tensorflow.org/forum/#!forum/tfjs-announce) mailing list.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs/4737)
<!-- Reviewable:end -->